move Isovist node into Visibility category and rename
